beninmarche/requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.5
Deprecated==1.2.14
Django==4.2.6
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-rest-swagger==0.3.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
multidict==6.0.4
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
twilio==8.10.0
typing_extensions==4.8.0
uritemplate==4.1.1
validate-email==1.3
wrapt==1.16.0
yarl==1.9.2
